se logro modificar todo lo necesario para que funcione los test de transaction

+++ src/main/java/com/unq/desa/criptoP2P/model/dto/ActiveCryptoReportDto.java
@@ -15,6 +15,6 @@ import java.util.List;
 public class ActiveCryptoReportDto {
     private LocalDateTime date;
     private Double dolarValue;
-    private Integer argValue;
+    private Double argValue;
     private List<ActivosDto> activosDto;
 }
